+++ dynsys/make_docs.py
@@ -50,6 +50,8 @@ make_html("damper")
 make_html("mesh")
 make_html("nodle")
 make_html("wind")
+#%%
+make_html("wind_section")
 
 #%%
 # List markdown files in root folder to be rendered